+++ lib/api/index.js
@@ -1,0 +1,5 @@
+/**
+ * @module  api
+ */
+
+module.exports = require('./context');
