+++ lib/api/random.js
@@ -1,139 +1,43 @@
-var debug = require('debug')('mongodb-datasets:schema:context'),
-  _ = require('underscore'),
-  bson = require('bson'),
-  chance = require('chance').Chance(),
+var chance = require('chance').Chance(),
   faker = require('faker'),
   helpers = require('./helpers');
-
-/**
- * Context object is solely used as context in function _.template
- * @class
- * @param {!Document} host The Document this Context is used
- */
-function Context (host) {
-  if (!(this instanceof Context)) return new Context(host);
-
-  /**
-   * Expose Chance.js library to clients
-   * @method
-   * @api public
-   * @see  {@link http://chancejs.com|Chance}
-   * @since 0.1.0
-   * @deprecated since 0.2.0
-   */
-  this.chance = chance;
-
-  /**
-   * Expose Faker.js library to clients
-   * @method
-   * @api public
-   * @see  {@link https://github.com/Marak/Faker.js|Faker}
-   * @since 0.1.0
-   * @deprecated since 0.2.0
-   */
-  this.faker = faker;
-
-  /**
-   * Internal state of this Context
-   * @member {Object}
-   * @private
-   * @property {boolean} display Visibility of the current Field
-   * @property {*} override Value overriding the default `_.template` output
-   */
-  this._state = {
-    display: undefined, // when true, the current field will not be visible
-    override: undefined // if present, used to override the template output
-  };
-
-  /**
-   * Collection of utility methods for clients
-   * @member {Object}
-   * @api public
-   * @property {function} sample {@link http://underscorejs.org/#sample}
-   * @since  0.1.0
-   */
-  this.util = {
-    sample: _.sample
-  };
-
-  /**
-   * Counter
-   * @method
-   * @api public
-   * @param  {number} [id=0]    Index of the counter to increment
-   * @param  {number} [start=0] Initial value,
-   *                            valid when the counter is initialized
-   * @param  {number} [step=1]  Increment of each count
-   * @return {number}           Current value of the specified counter
-   * @since  0.1.0
-   */
-  this.counter = function (id, start, step) {
-    id = id || 0; // though id=0 is false, does not matter
-    debug('counter called id=%d', id);
-    var counter = host._schema._state.counter; //pointer
-    if (typeof counter[id] === 'undefined') {
-      return (counter[id] = start || 0);
-    }
-    return (counter[id] += (step || 1));
-  };
-
-  /**
-   * Total number of generated Document's in the current run
-   * @instance
-   * @api public
-   * @memberOf Context
-   * @member {number} _$size
-   * @since  0.1.4
-   */
-  Object.defineProperty(this, '_$size', {
-    get: function () {
-      return host._schema._state.size;
-    }
-  });
-
-  /**
-   * Index of the current Document being generated, starting from 0
-   * @instance
-   * @api public
-   * @memberOf Context
-   * @member {number} _$index
-   * @since  0.1.4
-   */
-  Object.defineProperty(this, '_$index', {
-    get: function () {
-      return host._schema._state.index;
-    }
-  });
-
-}
-
-/**
- * Change the visibility of current Field
- * @method
- * @api public
- * @param  {boolean} pred Effective if true
- * @since  0.1.4
- */
-Context.prototype.hide = function (pred) {
-  if (pred) this._state.display = false;
-};
-
-/**
- * Convenient method to reset the internal state of this Context
- * @method
- * @private
- * @param  {boolean} hard True when the current Document finishes generating
- */
-Context.prototype._reset = function (hard) {
-  this._state.override = undefined;
-  if (hard) {
-    this._state.display = undefined;
-  }
-};
 
 /////////////////////////////////
 // new API for random function //
 /////////////////////////////////
+
+// function APIRandom () {
+//   if (!(this instanceof APIRandom)) return new APIRandom();
+// }
+
+// @todo how to isolate the random function API.
+
+/**
+ * [APIRandom description]
+ * @class
+ * @type {Object}
+ */
+var APIRandom = {};
+
+/**
+ * Expose Chance.js library to clients
+ * @method
+ * @api public
+ * @see  {@link http://chancejs.com|Chance}
+ * @since 0.1.0
+ * @deprecated since 0.2.0
+ */
+APIRandom.chance = chance;
+
+/**
+ * Expose Faker.js library to clients
+ * @method
+ * @api public
+ * @see  {@link https://github.com/Marak/Faker.js|Faker}
+ * @since 0.1.0
+ * @deprecated since 0.2.0
+ */
+APIRandom.faker = faker;
 
 // tile 1: heavy aggregation
 
@@ -141,7 +45,7 @@ Context.prototype._reset = function (hard) {
  * Generate a random number (floating point, or integer)
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {boolean} integer Produce an integer. True by default.
  * @property {boolean} float Produce a floating point number.
  * @property {number} min Lower range of result
@@ -165,7 +69,7 @@ Context.prototype._reset = function (hard) {
  * // returns, for example, 19
  * num({age: 'teen'});
  */
-Context.prototype.num = function () {
+APIRandom.num = function () {
   var opts = helpers.mix(arguments);
   var mm = opts.min !== undefined ? {min: Number(opts.min)} : {};
   var fix = opts.fixed !== undefined ? {fixed: Number(opts.fixed)} : {};
@@ -185,7 +89,7 @@ Context.prototype.num = function () {
  * By default, generate a name without prefix or middle name.
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {string} gender Gender: `male`/`female`
  * @property {boolean} first Only first name
  * @property {boolean} last Only last name
@@ -208,7 +112,7 @@ Context.prototype.num = function () {
  * // returns, for example, "Mr. Bruce Dale Morrison"
  * name('prefix', {gender: 'male', middle: 'full'});
  */
-Context.prototype.name = function () {
+APIRandom.name = function () {
   var opts = helpers.mix(arguments);
   var g = opts.gender ? {gender: opts.gender} : {};
   if (opts.first) return chance.first(g);
@@ -220,7 +124,7 @@ Context.prototype.name = function () {
     p = helpers.mix(g, opts.prefix === 'full' ? {full: true} : {});
     prefix = chance.prefix(p) + ' ';
   }
-  name = chance.name(_.extend(g, m));
+  name = chance.name(helpers.mix(g, m));
   return prefix + name;
 };
 
@@ -229,7 +133,7 @@ Context.prototype.name = function () {
  * By default, generate a random street address.
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {boolean} array Format the output as an array?
  * @property {string} delimiter Used to join components to a single string.
  *                              Default is " ".
@@ -257,7 +161,7 @@ Context.prototype.name = function () {
  * // returns, for example, [ "32 Karki Street", "Keiduciv" ]
  * address('array', 'city', 'street');
  */
-Context.prototype.address = function () {
+APIRandom.address = function () {
   var opts = helpers.mix(arguments);
   if (opts.full) opts = helpers.mix('street', 'city', 'state', 'zip', opts);
   var dl = opts.delimiter ? '' + opts.delimiter : ', ';
@@ -285,7 +189,7 @@ Context.prototype.address = function () {
  * Generate a random date
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {(boolean|string)} string Return a string instead with format of:
  *           + `true` - short format, i.e. "mm/dd/yyyy"
  *           + `long` - `Date.prototype.toDateString()`, e.g. "Tue Dec 30 2014"
@@ -304,7 +208,7 @@ Context.prototype.address = function () {
  * @return {(Date|string)}
  * @since  0.2.0
  */
-Context.prototype.date = function () {
+APIRandom.date = function () {
   var opts = helpers.mix(arguments);
   var d;
   if (opts.min || opts.max) {
@@ -343,7 +247,7 @@ Context.prototype.date = function () {
  * By default, generate a random one-char string.
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {number} length Default is 1
  * @property {string} pool Default is all alphanum and '!@#$%^&*()'
  * @property {string} casing Options are `lower` and `upper`
@@ -357,7 +261,7 @@ Context.prototype.date = function () {
  * // returns, for example, "g6dnc4rqhi6h"
  * str('hash', {length: 12});
  */
-Context.prototype.str = function () {
+APIRandom.str = function () {
   var opts = helpers.mix(arguments);
   var p = {}, c = opts.casing ? {casing: opts.casing} : {};
   p = opts.symbol ? {symbols: true} : p;
@@ -381,7 +285,7 @@ Context.prototype.str = function () {
  * By default, generate a string of random credit card number.
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {boolean} array Format the output as an array?
  * @property {string} delimiter Used to join components to a single string.
  *                              Default is " ".
@@ -403,7 +307,7 @@ Context.prototype.str = function () {
  * // returns, for example, [ 'visa', '4174049365395374', '01/2023' ]
  * cc('array', 'type', 'num', 'exp');
  */
-Context.prototype.cc = function () {
+APIRandom.cc = function () {
   var opts = helpers.mix(arguments);
   var rtn = [], t = chance.cc_type({raw: true});
   if (opts.type) {
@@ -424,7 +328,7 @@ Context.prototype.cc = function () {
  * By default, generate one `lorem`-like sentence.
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {number} sentence Generate a paragraph of num sentences
  * @property {number} word Generate a sentence of num words
  * @property {number} syllable Generate a word with num syllables
@@ -436,7 +340,7 @@ Context.prototype.cc = function () {
  * // returns, for example, "fatelu"
  * text({syllable: 3});
  */
-Context.prototype.text = function () {
+APIRandom.text = function () {
   var opts = helpers.mix(arguments);
   if (opts.sentence)
     return chance.paragraph({sentences: Number(opts.sentence)});
@@ -453,7 +357,7 @@ Context.prototype.text = function () {
  * Generate a random ID for different purposes.
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {boolean} guid Generate a guid (default behavior)
  * @property {boolean} fb Generate a facebook ID
  * @property {boolean} twitter Generate a twitter ID
@@ -469,7 +373,7 @@ Context.prototype.text = function () {
  * // returns, for example, '1000052961667394'
  * id('fb');
  */
-Context.prototype.id = function () {
+APIRandom.id = function () {
   var opts = helpers.mix(arguments);
   if (opts.fb) return chance.fbid();
   if (opts.twitter) return chance.twitter();
@@ -484,7 +388,7 @@ Context.prototype.id = function () {
  * By default, generate "latitude, longitude".
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {boolean} array Format the output as an array?
  * @property {string} delimiter Used to join components to a single string.
  *                              Default is " ".
@@ -502,7 +406,7 @@ Context.prototype.id = function () {
  * // returns, for example, -75.1
  * coordinates('lat', {fixed: 1});
  */
-Context.prototype.coordinates = function () {
+APIRandom.coordinates = function () {
   var opts = helpers.mix(arguments);
   var dl = opts.delimiter ? opts.delimiter : ', ';
   var f = opts.fixed === undefined ? {} : {fixed: Number(opts.fixed)};
@@ -526,7 +430,7 @@ Context.prototype.coordinates = function () {
  * @return {string}
  * @since  0.2.0
  */
-Context.prototype.random = function (pool) {
+APIRandom.random = function (pool) {
   // var opts = helpers.mix(pool);
   switch (pool) {
     case 'tld':
@@ -554,7 +458,7 @@ Context.prototype.random = function (pool) {
  * Generate a random phone number
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {boolean} areacode Only return areacode, e.g. "(510)"
  * @return {string}
  * @since  0.2.0
@@ -563,7 +467,7 @@ Context.prototype.random = function (pool) {
  * // returns, for example, '(622) 941-1923'
  * phone();
  */
-Context.prototype.phone = function () {
+APIRandom.phone = function () {
   var opts = helpers.mix(arguments);
   if (opts.areacode) return chance.areacode();
   return chance.phone();
@@ -573,7 +477,7 @@ Context.prototype.phone = function () {
  * Generate a random IP address
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {boolean} v6 Use IPv6
  * @return {string}
  * @since  0.2.0
@@ -582,7 +486,7 @@ Context.prototype.phone = function () {
  * // returns, for example, '47.55.14.224'
  * ip();
  */
-Context.prototype.ip = function () {
+APIRandom.ip = function () {
   var opts = helpers.mix(arguments);
   if (opts.v6) return chance.ipv6();
   return chance.ip();
@@ -594,12 +498,12 @@ Context.prototype.ip = function () {
  * Generate a random domain
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {string} tld With a fixed a top level domain
  * @return {string}
  * @since  0.2.0
  */
-Context.prototype.domain = function () {
+APIRandom.domain = function () {
   return chance.domain(helpers.mix(arguments));
 };
 
@@ -607,12 +511,12 @@ Context.prototype.domain = function () {
  * Generate a random email
  * @method
  * @api public
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {string} domain With a fixed domain
  * @return {string}
  * @since  0.2.0
  */
-Context.prototype.email = function () {
+APIRandom.email = function () {
   return chance.email(helpers.mix(arguments));
 };
 
@@ -621,7 +525,7 @@ Context.prototype.email = function () {
  * @method
  * @api public
  * @see  {@link http://chancejs.com/#color}
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {string} format Format of color. Options are hex, shorthex, rgb
  * @property {boolean} greyscale Only grayscale color
  * @return {string}
@@ -631,7 +535,7 @@ Context.prototype.email = function () {
  * // returns, for example, '#d67118'
  * color({format: 'hex'});
  */
-Context.prototype.color = function () {
+APIRandom.color = function () {
   return chance.color(helpers.mix(arguments));
 };
 
@@ -640,121 +544,16 @@ Context.prototype.color = function () {
  * @method
  * @api public
  * @see  {@link http://chancejs.com/#ssn}
- * @param {...Context.OptionHash}
+ * @param {...APIRandom.OptionHash}
  * @property {boolean} ssnFour Only last four digits
  * @property {boolean} dashes With dashes?
  * @return {string}
  * @since  0.2.0
  */
-Context.prototype.ssn = function () {
+APIRandom.ssn = function () {
   var opts = helpers.mix(arguments);
   if (!opts.dashes) opts.dashes = false;
   return chance.ssn(opts);
 };
 
-/////////////////////////////////////////
-// Constructors of all supported types //
-/////////////////////////////////////////
-
-/**
- * Replace `_.template` ouput
- * @method
- * @api public
- * @param {} i Must be acceptable by `Number` constructor
- * @since  0.1.0
- */
-Context.prototype.Number = function (i) {
-  this._state.override = Number(i);
-};
-
-/**
- * Replace `_.template` ouput
- * @method
- * @api public
- * @param {} b Must be acceptable by `Boolean` constructor
- * @since  0.1.0
- */
-Context.prototype.Boolean = function (b) {
-  this._state.override = Boolean(b);
-};
-
-/**
- * Replace `_.template` ouput
- * @method
- * @api public
- * @param {} s Must be acceptable by `String` constructor
- * @since  0.1.0
- */
-Context.prototype.String = function (s) {
-  this._state.override = String(s);
-};
-
-/**
- * Replace `_.template` ouput
- * @method
- * @api public
- * @param {} d Must be acceptable by `Date` constructor
- * @since  0.1.0
- */
-Context.prototype.Date = function (d) {
-  this._state.override = new Date(d);
-};
-
-/**
- * Replace `_.template` ouput
- * @method
- * @api public
- * @param {} i Must be acceptable by `Long` constructor in bson package
- * @see  {@link https://github.com/mongodb/js-bson|bson}
- * @since  0.1.0
- */
-Context.prototype.NumberLong = function (i) {
-  this._state.override = new bson.Long(i);
-};
-
-/**
- * Replace `_.template` ouput
- * @method
- * @api public
- * @see  {@link https://github.com/mongodb/js-bson|bson}
- * @since  0.1.0
- */
-Context.prototype.MinKey = function () {
-  this._state.override = new bson.MinKey();
-};
-
-/**
- * Replace `_.template` ouput
- * @method
- * @api public
- * @see  {@link https://github.com/mongodb/js-bson|bson}
- * @since  0.1.0
- */
-Context.prototype.MaxKey = function () {
-  this._state.override = new bson.MaxKey();
-};
-
-/**
- * Replace `_.template` ouput
- * @method
- * @api public
- * @see  {@link https://github.com/mongodb/js-bson|bson}
- * @since  0.1.0
- */
-Context.prototype.Timestamp = function () {
-  this._state.override = new bson.Timestamp();
-};
-
-/**
- * Replace `_.template` ouput
- * @method
- * @api public
- * @param {} i Must be acceptable by `ObjectId` constructor in bson package
- * @see  {@link https://github.com/mongodb/js-bson|bson}
- * @since  0.1.0
- */
-Context.prototype.ObjectID = function (i) {
-  this._state.override = new bson.ObjectId(i);
-};
-
-module.exports = Context;
+module.exports = APIRandom;
