+++ lib/schema/array.js
@@ -8,11 +8,11 @@ var DEFAULTS = {
   size: [2, 5]
 };
 
-util.inherits(ArrayField, Entry);
 /**
- * Class representation of arrays in template schema
+ * Internal representation of Arrays in template schema
  * @class
  * @augments {Entry}
+ * @memberOf module:schema
  * @param {!Array} array   Raw user-supplied array
  * @param {string} name   Key corresponding to the field
  * @param {!Entry} parent  Immediate enclosing Entry
@@ -62,6 +62,7 @@ function ArrayField (array, name, parent) {
 
   debug('array <%s> build complete', name);
 }
+util.inherits(ArrayField, Entry);
 
 ArrayField.prototype.next = function () {
   var data = [];
