+++ lib/schema/document.js
@@ -1,24 +1,22 @@
 var util = require('util');
 var helpers = require('./helpers');
 var Entry = require('./entry');
-var Context = require('./context');
 var debug = require('debug')('mongodb-datasets:schema:document');
 
-// doc must be a non-array object
-util.inherits(Document, Entry);
 /**
- * [Document description]
+ * Internal representation of Objects (`{}`)
  * @class
  * @augments {Entry}
- * @param {object} document [description]
- * @param {string} name     [description]
- * @param {(Entry|Schema)} parent   [description]
+ * @memberOf module:schema
+ * @param {object} document 
+ * @param {string} name     
+ * @param {(Entry|Schema)} parent   
  */
 function Document (document, name, parent) {
   if (!(this instanceof Document)) return new Document(document, name, parent);
   debug('document <%s> being built', name);
   Entry.call(this, document, name, parent);
-  this._context = new Context(this);
+  this._context = helpers.getAPI(this);
   this._currVal = {}; // 'this' object in evaluating user expressions
   this._this = this._currVal;
   Object.defineProperty(this._currVal, '_$parent', {
@@ -45,6 +43,7 @@ function Document (document, name, parent) {
   }
   debug('document <%s> build complete', name);
 }
+util.inherits(Document, Entry);
 
 Document.prototype.next = function () {
   this._clean();
