+++ lib/schema/entry.js
@@ -1,12 +1,13 @@
-var debug = require('debug')('mongodb-datasets:entry');
+var debug = require('debug')('mongodb-datasets:schema:entry');
 
 /**
- * [Entry description]
+ * High level representation of all fields in template schemas
  * @class
  * @abstract
- * @param {*} content [description]
- * @param {string} name    [description]
- * @param {(Entry|Schema)} parent  [description]
+ * @memberOf module:schema
+ * @param {*} content
+ * @param {string} name   
+ * @param {(Entry|Schema)} parent 
  */
 function Entry (content, name, parent) {
   if (!(this instanceof Entry)) return new Entry(content, name, parent);
