+++ lib/schema/field.js
@@ -7,15 +7,14 @@ _.templateSettings = {
   interpolate: /\{\{(.+?)\}\}/g
 };
 
-// field must be neither an array nor an object
-util.inherits(Field, Entry);
 /**
- * [Field description]
+ * Internal representation of non-object non-array fields in template schema
  * @class
  * @augments {Entry}
- * @param {(string|number|boolean)} field  [description]
- * @param {string} name   [description]
- * @param {Entry} parent [description]
+ * @memberOf module:schema
+ * @param {(string|number|boolean)} field 
+ * @param {string} name  
+ * @param {Entry} parent
  */
 function Field (field, name, parent) {
   if (!(this instanceof Field)) return new Field(field, name, parent);
@@ -33,6 +32,7 @@ function Field (field, name, parent) {
     this._compiled = _.template(this._content);
   }
 }
+util.inherits(Field, Entry);
 
 Field.prototype.next = function () {
   if (typeof this._currVal !== 'undefined') {
