+++ lib/schema/index.js
@@ -1,11 +1,18 @@
+// @todo wrong way to mark class as a member of a module
+
+/**
+ * @module  schema
+ */
+
 var helpers = require('./helpers'),
   debug = require('debug')('mongodb-datasets:schema:index');
 
 /**
- * [Schema description]
+ * Internal representation of a user-defined schema
  * @class
- * @param {object} sc   [description]
- * @param {number} size [description]
+ * @memberOf module:schema
+ * @param {object} sc  
+ * @param {number} size
  */
 function Schema (sc, size) {
   if (!(this instanceof Schema)) return new Schema(sc);
